+++ code/src/cljs/sixsq/nuvla/ui/profile/events.cljs
@@ -477,6 +477,7 @@
   (fn [_ [_ response]]
     {:fx [[:dispatch [::set-error (-> response :response :message)]]]}))
 
+
 (reg-event-fx
   ::code-validation-2fa-success
   (fn [_ [_ success-header success-content]]
@@ -489,60 +490,102 @@
 
 
 (reg-event-fx
-  ::validate-2fa-activation
-  (fn [{{:keys [::spec/callback-2fa]} :db} [_ token success-header success-content]]
+  ::two-factor-enabled
+  (fn [_ [_ success-header success-content]]
+    {:fx [[:dispatch [::messages-events/add
+                      {:header  success-header
+                       :content success-content
+                       :type    :success}]]
+          [:dispatch [::get-user]]
+          [:dispatch [::set-two-factor-step :save-secret]]]}))
+
+
+(reg-event-fx
+  ::two-factor-disabled
+  (fn [_ [_ success-header success-content]]
+    {:fx [[:dispatch [::messages-events/add
+                      {:header  success-header
+                       :content success-content
+                       :type    :success}]]
+          [:dispatch [::get-user]]
+          [:dispatch [::close-modal]]]}))
+
+
+(reg-event-fx
+  ::two-factor-auth-callback-exec
+  (fn [{{:keys [::spec/two-factor-callback]} :db} [_ token success-dispatch-vec]]
     {:http-xhrio {:method          :put
-                  :uri             callback-2fa
+                  :uri             two-factor-callback
                   :format          (ajax/json-request-format)
                   :params          {:token token}
                   :response-format (ajax/json-response-format {:keywords? true})
-                  :on-success      [::code-validation-2fa-success success-header success-content]
+                  :on-success      success-dispatch-vec
                   :on-failure      [::code-validation-2fa-failed]}}))
 
 
 (reg-event-db
-  ::set-callback-2fa
-  (fn [db [_ response loading-key]]
-    (let [callback-url (:location response)]
-      (-> db
-          (update ::spec/loading disj loading-key)
-          (assoc ::spec/callback-2fa callback-url)))))
+  ::set-two-factor-step
+  (fn [db [_ step]]
+    (assoc db ::spec/two-factor-step step)))
 
 
 (reg-event-fx
-  ::enable-2fa
-  (fn [{{:keys [::spec/user] :as db} :db}]
-    {:db                     (-> db
-                                 (update ::spec/loading conj :enable-2fa)
-                                 (assoc ::spec/open-modal :enable-2fa)
-                                 (assoc ::spec/callback-2fa nil))
-     ::cimi-api-fx/operation [(:id user) "enable-2fa"
-                              #(if (instance? js/Error %)
-                                 (let [{:keys [status message]} (response/parse-ex-info %)]
-                                   (dispatch [::set-callback-2fa nil :enable-2fa])
-                                   (dispatch [::messages-events/add
-                                              {:header  (cond-> (str "Enable 2FA failed!")
-                                                                status (str " (" status ")"))
-                                               :content message
-                                               :type    :error}]))
-                                 (dispatch [::set-callback-2fa %1 :enable-2fa]))
-                              {:method "email"}]}))
+  ::set-two-factor-op-response
+  (fn [{{:keys [::spec/two-factor-enable?] :as db} :db} [_ {:keys [location secret] :as _response}]]
+    (let [next-step (if secret :totp :email)]
+      {:db (-> db
+               (update ::spec/loading disj :two-factor-auth)
+               (assoc ::spec/two-factor-callback location)
+               (assoc ::spec/two-factor-secret secret))
+       :fx [(when two-factor-enable?
+              [:dispatch [::set-two-factor-step next-step]])]})))
 
 
 (reg-event-fx
-  ::disable-2fa
+  ::two-factor-operation-call
+  (fn [{{:keys [::spec/user
+                ::spec/two-factor-enable?
+                ::spec/two-factor-method] :as db} :db}]
+    (let [op             (if two-factor-enable? "enable-2fa" "disable-2fa")
+          enable-disable (if two-factor-enable? "Enable" "Disable")]
+      {:db (update db ::spec/loading conj :two-factor-auth)
+       ::cimi-api-fx/operation
+       [(:id user) op
+        #(if (instance? js/Error %)
+           (let [{:keys [status message]} (response/parse-ex-info %)]
+             (dispatch [::set-two-factor-op-response nil])
+             (dispatch [::close-modal])
+             (dispatch [::messages-events/add
+                        {:header  (cond-> (str enable-disable " 2FA failed!")
+                                          status (str " (" status ")"))
+                         :content message
+                         :type    :error}]))
+           (dispatch [::set-two-factor-op-response %1]))
+        {:method two-factor-method}]})))
+
+
+(reg-event-fx
+  ::select-method
+  (fn [{db :db} [_ method]]
+    {:db (assoc db ::spec/two-factor-method method)
+     :fx [[:dispatch [::two-factor-operation-call]]]}))
+
+
+(reg-event-db
+  ::two-factor-enable
+  (fn [db]
+    (-> db
+        (assoc ::spec/open-modal :two-factor-auth)
+        (assoc ::spec/two-factor-step :select-method)
+        (assoc ::spec/two-factor-enable? true))))
+
+
+(reg-event-fx
+  ::two-factor-disable
   (fn [{{:keys [::spec/user] :as db} :db}]
-    {:db                     (-> db
-                                 (update ::spec/loading conj :disable-2fa)
-                                 (assoc ::spec/open-modal :disable-2fa)
-                                 (assoc ::spec/deactivation-2fa-callback nil))
-     ::cimi-api-fx/operation [(:id user) "disable-2fa"
-                              #(if (instance? js/Error %)
-                                 (let [{:keys [status message]} (response/parse-ex-info %)]
-                                   (dispatch [::set-callback-2fa nil :disable-2fa])
-                                   (dispatch [::messages-events/add
-                                              {:header  (cond-> (str "Disable 2FA failed!")
-                                                                status (str " (" status ")"))
-                                               :content message
-                                               :type    :error}]))
-                                 (dispatch [::set-callback-2fa %1 :disable-2fa]))]}))
+    {:db (-> db
+             (assoc ::spec/open-modal :two-factor-auth)
+             (assoc ::spec/two-factor-enable? false)
+             (assoc ::spec/two-factor-step :disable)
+             (assoc ::spec/two-factor-method (:auth-method-2fa user)))
+     :fx [[:dispatch [::two-factor-operation-call]]]}))